pyproject.toml
[tool.poetry]
name = "ytplaylistsync"
version = "0.1.0"
description = ""
authors = ["MrXaxen <mrxaxen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^2.42.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
argparse = "^1.4.0"
google-auth = "^2.6.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
